+++ website/src/routes/api/(actions)/normalize/properties.ts
@@ -1,0 +1,31 @@
+import type { PropertyProps } from '~/components';
+
+export const properties: Record<string, PropertyProps> = {
+  TForm: {
+    modifier: 'extends',
+    type: {
+      type: 'union',
+      options: [
+        {
+          type: 'custom',
+          name: 'NormalizeForm',
+          href: '../NormalizeForm/',
+        },
+        'undefined',
+      ],
+    },
+  },
+  form: {
+    type: {
+      type: 'custom',
+      name: 'TForm',
+    },
+  },
+  Action: {
+    type: {
+      type: 'custom',
+      name: 'NormalizeAction',
+      href: '../NormalizeAction/',
+    },
+  },
+};